refactor: move favorites and movies for own directories
