+++ lib/domain/interfaces/models/media.dart
@@ -1,5 +1,3 @@
-import 'dart:convert';
-
 abstract class Media {
   final int id;
   final String title;
