+++ lib/domain/interfaces/models/series/popular_series.dart
@@ -1,5 +1,3 @@
-import 'dart:convert';
-
 import '../media.dart';
 import 'package:json_annotation/json_annotation.dart';
