+++ test/widgets/media_card_test.dart
@@ -1,4 +1,4 @@
-import 'package:dart_plus_app/domain/interfaces/models/popular_movies.dart';
+import 'package:dart_plus_app/domain/interfaces/models/movies/popular_movies.dart';
 import 'package:dart_plus_app/presentation/widgets/media_card.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
